+++ app/src/main/java/com/example/findu/domain/model/my/MyReportHistoryData.kt
@@ -2,13 +2,13 @@ package com.example.findu.domain.model.my
 
 data class MyReportHistoryData(
     val isLast: Boolean,
-    val lastReportId: Int,
+    val lastReportId: Long,
     val reports: List<Report>
 ) {
     data class Report(
         val date: String,
         val location: String,
-        val reportId: Int,
+        val reportId: Long,
         val tag: String,
         val thumbnailImageUrl: String,
         val title: String
